Mensaje de confirmación para elminar un Espectáculo

+++ js/shows.js
@@ -85,9 +85,9 @@ jQuery(function($) {
 Gets shows list to display
 */
 jQuery(function($) {
-  getShowList()
+  getShowList();
 });
-function getShowList(){
+function getShowList() {
   $.ajax({
     type: "GET",
     url: "controller/controller-list-shows.php",
@@ -121,7 +121,8 @@ function fillTableShows(shows) {
         '" type="button" class="btn btn-primary btn-lg update-show"><i class="mdi mdi-pencil mdi-24px"></i></button>' +
         '<button show-id="' +
         show.id_show +
-        '" type="button" class="btn btn-primary btn-lg delete-show"><i class="mdi mdi-delete mdi-24px"></i></button>' +
+        '" type="button" class="btn btn-primary btn-lg delete-show" >' +
+        '<i class="mdi mdi-delete mdi-24px"></i></button>' +
         " </td>";
 
       tr += "<tr>" + information + "</tr>";
@@ -144,8 +145,12 @@ jQuery(function($) {
 jQuery(function($) {
   $("#table-shows tbody").on("click", "button.delete-show", function() {
     const idShow = $(this).attr("show-id");
-    //deleteImage(idShow);
-    deleteShow(idShow);
+    conf = confirm("El espectáculo será eliminado. ¿Desea continuar?");
+    if (conf) {
+      deleteShow(idShow);
+    } else {
+      return false;
+    }
   });
 });
 
@@ -153,7 +158,6 @@ jQuery(function($) {
   Gets all the values from the form
 */
 jQuery(function($) {
-
   $("#save-new-show").click(function() {
     if (validateRequiredFileds(".form-add-newshow")) {
       let genres = [];
@@ -170,11 +174,11 @@ jQuery(function($) {
       ) {
         const showData = {
           artist: $("#nameShow").val(),
-          amount:$("#money").val(),
+          amount: $("#money").val(),
           genres: genres,
           datesTime: dateTimes,
-          imgMobile: $('#img-mobile')[0].files[0].name,
-          imgDesktop: $("#img-desktop")[0].files[0].name,
+          imgMobile: $("#img-mobile")[0].files[0].name,
+          imgDesktop: $("#img-desktop")[0].files[0].name
         };
 
         $.ajax({
@@ -184,23 +188,22 @@ jQuery(function($) {
           contentType: "application/json; charset=utf-8",
           dataType: "json",
           success: function(data) {
-            sendImages()
+            sendImages();
             resetForm();
           },
           error: function(errMsg) {
             console.error(errMsg.responseJSON.error);
           }
         });
-
       }
     }
   });
 });
 
-function sendImages(){
-  const form_data = new FormData();                  
-  form_data.append('img-desktop', $("#img-desktop").prop('files')[0]);
-  form_data.append('img-mobile', $("#img-mobile").prop('files')[0]);
+function sendImages() {
+  const form_data = new FormData();
+  form_data.append("img-desktop", $("#img-desktop").prop("files")[0]);
+  form_data.append("img-mobile", $("#img-mobile").prop("files")[0]);
   $.ajax({
     type: "POST",
     url: "controller/controller-create-shows-images.php",
@@ -209,9 +212,9 @@ function sendImages(){
     contentType: false,
     processData: false,
     success: function() {
-      notifications ("Espectáculo creado.", 'success')
+      notifications("Espectáculo creado.", "success");
       $("#add-newShow").hide();
-      getShowList()
+      getShowList();
       $("#list-show").show();
     },
     error: function(errMsg) {
@@ -220,32 +223,31 @@ function sendImages(){
   });
 }
 
-function resetForm(){
-  $('.form-add-newshow').trigger("reset");
+function resetForm() {
+  $(".form-add-newshow").trigger("reset");
   dateTimes = [];
   $("#date-timeShowList").empty();
-  $('#multiple-checkboxes').multiselect('updateButtonText');
+  $("#multiple-checkboxes").multiselect("updateButtonText");
 }
 
-function deleteShow(idShow){
-   
-    const showData = {
-      idShow: idShow
-    };
-    $.ajax({
-      type: "POST",
-      url: "controller/controller-deleteShow.php",
-      data: JSON.stringify(showData),
-      contentType: "application/json; charset=utf-8",
-      dataType: "json",
-      success: function(showData) {
-        console.log(showData);
-        notifications ("Espectaculo eliminado exitosamente.", 'success')
-        getShowList()
-      },
-      error: function(errMsg) {
+function deleteShow(idShow) {
+  const showData = {
+    idShow: idShow
+  };
+  $.ajax({
+    type: "POST",
+    url: "controller/controller-deleteShow.php",
+    data: JSON.stringify(showData),
+    contentType: "application/json; charset=utf-8",
+    dataType: "json",
+    success: function(showData) {
+      console.log(showData);
+      notifications("Espectaculo eliminado exitosamente.", "success");
+      getShowList();
+    },
+    error: function(errMsg) {
       console.log(errMsg);
-        notifications ("Error al eliminar un espectaculo.", 'error')
-      }
-    });
-  }
+      notifications("Error al eliminar un espectaculo.", "error");
+    }
+  });
+}
